requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-whois
rapidfuzz
pandas
streamlit
dnspython
validators
